Create searchInsert.cpp
You are given a sorted array 'arr' of distinct values and a target value 'm. You need to search for the index of the target value in the array.
If the value is present in the array, then return its index.
If the value is not present, determine the index where it would be inserted in the array while maintaining the sorted order.
Example:
Input:
arr = (1, 2, 4, 71, m = 6
Output: 3
Explanation; If the given array arr' is: 11, 2, 4, 7] and m - 6. We insert
m = 6 in the array and get
'arr' as:
[1, 2, 4, 6, 71. The position of 6 is 3 (according to 0-based indexing)